refactor: fix clippy warn
